api_segura/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Authlib==1.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
pytest==7.4.0
pytest-flask>=1.3.0